# [Minor] Refactor `parse` Function