Refactor `parse` function if all the characters in a base32 encoded ULID are part of the expected base32 character set into its own function.

Since the if condition checks if one of the entries is `0xFF` and then returns true for `ErrInvalidCharacters` I refactored the break into a for loop that exits true early if the entry is `0xFF`