@peterbourgon 2023-12-28
No objection, but I'll observe that the existing code will short-circuit in the `if` expr, so I _think_ this change should be net-zero. But to be sure, please add a benchmark :)