@peterbourgon 2023-12-28
Am I reading it correctly that this change is 22% more CPU and 18% less throughput i.e. strictly worse?