@el10savio 2023-12-29
Yeah, looks like this refactor made it less performant. 
Will close the PR to keep the current changes, thanks for the review @peterbourgon